requirements.txt
PyNaCl~=1.5.0